fix trace dataloader
